move wal creation to a method
